Config options added for fitting map to bounds
Three config options added to aid with fitting the plotted points on a
map to the map bounds.

- fitMapToBounds: Boolean (Default: true) - When true that map will fit
to the bounds of the Latitude and Longitude fields of the records
provided to the StoreMap plugin.

- minimumMapZoom: Boolean (Default: true) - When true the map will have
a minimum zoom level as defined by the minimumZoomLevel config option.

- minimumZoomLevel: Integer (Default: 16) - When minimumMapZoom is set
to true this config option can be used to define the integer that
represents zoom in Google Maps.