Info Window functionality
Ability to create info windows using Google Map Info Window Api added.

Various config options added to handle this new functionality.

markerTapInfoWindow: True if you wish markers to show Google Maps Info
Window. contentString config can be used to define to the content of
Info Window.  If false the default 'markertap' event will fire.

contentString: The content string to be displayed within the info
window. If a string is set then it will be used as the content, if a
Function is used it will have the record passed in.

closeInfoWindowMapTap: False if you want Info Windows to stay open on a
map click event.  If false the default 'mapclick' event will fire.  If
True all Info Windows will close on a map click.